Move instance variable logic to separate module
